code duplication removal
